initial test fo update deps
